Add xdp-forward utility
Add the xdp-forward utility, the beginning of an XDP forwarding plane utility.
This is a port of the xdp_fwd utility from samples/bpf in the kernel tree, with
the userspace component converted to the xdp-tools command structure (with
'load' and 'unload' commands), and using libxdp to load and unload.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>